Add new type 'pbenum'
Specifying `type:"pbenum"` in a tag will cause `kong` to use the pbenum mapper.

The pbenum mapper will attempt to load the enum descriptor via the `EnumDescriptor()`
method that should be generated as part of a `protoc` run.

If this method exists and we get back a valid FileDescriptor, the mapper will
attempt to lookup all of the available enums NAMES and allow you to use the
string name (as opposed to having to use the int32 value that an enum is set to).

We need this functionality because we generate our CLI opts from compiled protos
and make significant use of `enum`. Having to specify `--flag=10` (or another
int) is unpleasant.

I stuck pretty much all of the functionality in the mapper because I didn't want
to touch and complicate the existing `enum`, but maybe there's a better way to
do this?

Also - it would be great if the available pbenums could be exposed in help
via a var (such as `$pbenum`, same as `$enum`), but I'm not sure where such
functionality should live.