Export tag.items -> tag.Items
This makes item public, so that consumers can see their own tags. This
helps (e.g.) building completion generators that want to use their
custom tags on the struct.

Taking a *kong.Node as n, and:
~~~ go
for i, f := range n.Flags {
        fmt.Printf("%d %v\n", i, f.Tag.Items["completion"])
}
~~~

shows the value of the (custom) completion tag.

Signed-off-by: Miek Gieben <miek@miek.nl>