fix(Context.Run): Don't panic on unselected root node
Fixes regression introduced in 2544d3f00855f6857cf121ce04608296b5858e4a.

That commit refactored control flow slightly and missed a case
where the selected node is an application node but it does not
have a Run method.

The fix was straightforward:
Treat the "if application node" code path as a possible way to fill
`node`, and check it again afterwards.

Resolves #483